［需求／BUG］: 完成所有基础功能 ［模块］: ［说明］: ［作者］: 桥泸007

+++ tlias-ai-project/src/main/java/com/qiaolu/service/iml/DeptServiceIml.java
@@ -1,10 +1,13 @@
 package com.qiaolu.service.iml;
 
+import com.qiaolu.exception.BusinessException;
 import com.qiaolu.mapper.DeptMapper;
+import com.qiaolu.mapper.EmpMapper;
 import com.qiaolu.pojo.Dept;
 import com.qiaolu.service.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -14,6 +17,8 @@ public class DeptServiceIml implements DeptService {
 
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
 
     @Override
     public List<Dept> findAll() {
@@ -21,8 +26,13 @@ public class DeptServiceIml implements DeptService {
     }
 
     @Override
-    public void deleteById(Integer id) {
-        deptMapper.deleteById(id);
+    @Transactional(rollbackFor = BusinessException.class)
+    public void deleteById(Integer id) throws Exception {
+        if (empMapper.getNumByDeptID(id) == 0) {
+            deptMapper.deleteById(id);
+        } else {
+            throw new BusinessException("无法删除部门，因为该部门下存在员工");
+        }
     }
 
     @Override
